update transaction - removed previous txn and use latest txn feature
