+++ lib/intergallactic.js
@@ -7,7 +7,6 @@ const Account = require('./intergallactic/account'),
   Contract = require('./intergallactic/contract'),
   Gallactic = require('./intergallactic/gallactic'),
   Transaction = require('./intergallactic/transaction'),
-  TransactionV2 = require('./intergallactic/transaction.v2'),
   pckg = require('../package');
 
 /**
@@ -24,7 +23,6 @@ class Intergallactic {
     this.account = new Account(this.conn);
     this.gallactic = new Gallactic(this.conn);
     this.Transaction = Transaction.bind(null, this);
-    this.TransactionV2 = TransactionV2.bind(null, this);
     this.Contract = Contract.bind(null, this);
     this.utils = getUtil();
     this.version = getVersion();
