+++ lib/intergallactic/contract.js
@@ -76,7 +76,7 @@ class Contract {
       data: this._byteCode || contract.byteCode
     };
 
-    const callTx = new this.igc.TransactionV2(newTx);
+    const callTx = new this.igc.Transaction(newTx);
     return callTx.call(privateKey)
       .then(res => {
         if (res && res.body && res.body.result) {
