+++ lib/intergallactic/contract.method.js
@@ -36,7 +36,7 @@ class ContractMethod {
       data: this.encodeABI()
     };
 
-    const callTx = new this.igc.TransactionV2(newTx);
+    const callTx = new this.igc.Transaction(newTx);
     return callTx.call(privateKey)
       .then(res => {
         if (res && res.body && res.body.result) {
