+++ lib/intergallactic/transaction.js
@@ -1,4 +1,5 @@
-'use strict'
+'use strict';
+
 const config = require('../config/constant'),
   conversion = require('../utils/conversion'),
   gkeys = require('gallactickeys'),
@@ -21,402 +22,385 @@ const config = require('../config/constant'),
  * @example
  * const privKey = '<insert_private_key>'
  * const igc = new Intergallactic({ url: 'localhost:1337/rpc', protocol: 'jsonrpc' });
- * const txn = {
- *   from: [{
- *     address: testAcc.address,
- *     amount: 10,
- *   }],
- *   to: [{
- *     address: 'acWmVcNrHzxBF8L25vdiKLsz664ZGkYmPRj',
- *     amount: 10
- *   }]
- * };
- * const opt = { txnType: 1 // required, unit: 'boson' // default if not specified }
- * const newTxn = new igc.Transaction(txn, opt);
- * newTxn.signNBroadcast(privKey)
+ * const newTxn = new igc.Transaction({
+ *   to: 'acWmVcNrHzxBF8L25vdiKLsz664ZGkYmPRj'
+ *   amount: 100
+ * });
+ * newTxn.send(privKey)
  *  .then(res => res contains transaction info)
  */
 class Transaction {
-  constructor(igc, data = {}, opt = {}) {
+  constructor (igc, data = {}) {
+    if ([typeof data.to, typeof data.amount].includes('undefined')) {
+      throw new Error('Invalid or empty "to" and "amount" field. Please provide valid "to" and "amount" field')
+    }
     this.igc = igc;
-    if (!opt.type) {
-      throw new Error('Instantiate transaction require type value to decide whether send or call transaction.')
-    }
     this.conn = this.igc.conn;
-    this.utils = this.igc.utils;
-    this._txn = data;
-    this.txnType = getTxnSign(opt.type);
-    this.txnUnit = opt.unit || config.gallactic.transaction.defaultUnit;
-    this.chainId = opt.chainId;
-    this.sequence = opt.sequence;
+    this._to = data.to;
+    this._amount = data.amount;
+    this._unit = this._unit || config.gallactic.transaction.defaultUnit;
+
+    // general optional fields
+    if (data.type) this._typeId = data.type;
+    if (data.unit) this._unit = data.unit;
+    if (data.chainId) this._chainId = data.chainId;
+    if (data.sequence) this._sequence = data.sequence;
+
+    // assigning public var
+    if (this._typeId) this._type = this._getTxnType();
+
+    // special field for call
+    if (data.gasLimit) this._gasLimit = data.gasLimit;
+    if (data.data) this._data = data.data;
+
+    // special field for bond
+    if (data.publicKey) this._publicKey = data.publicKey;
+
+    // special field for permission
+    if (data.permissions) this._permissions = data.permissions;
+    if (data.set) this._set = data.set;
+
+    if (this._chainId && this._sequence) {
+      // TODO: add synchronous sign flag here?
+    }
   }
 
-  /**
-   * To sign transaction "synchronously" provided sequence and chainId upon instatiate
-   * @param {String} privKey private key string
-   * @returns {String} a transaction signature string
-   */
-  signSync(privKey) {
-    if (!this.chainId || typeof this.sequence !== 'number') {
-      throw new Error('Chain id or sequence is not defined. Synchronous sign require chainId and sequence as parameter upon instantiate')
+  sign (privKey) {
+    // condition if type is not exist
+    if (this._typeId) this.type = this._getTxnType();
+    if (!privKey) {
+      throw new Error('Signing transaction require "private key" value. Please assign private key upon signing');
     }
-
-    this.txn = buildTxnByType(this._txn, this.txnType, { unit: this.txnUnit });
-    // Set Transaction sequence by transaction type
-    setTxnSequence(this.txn, this.txnType, this.sequence);
-
-    return signByPrivKey(privKey, this.txn, this.txnType, this.chainId);
+    if (!this._type) {
+      throw new Error('Signing transaction require transaction "type" field. Please use declare instance with "type" field and try again');
+    }
+    this._setFrom(privKey);
+    return this._setChainId() // get chain id
+      .then(() => { return this._setSequence() }) // get sequence
+      .then(() => { return this._buildTxnByType() }) // build txn
+      .then(() => { return this._setSignatories(privKey) }) // sign
+      .then(() => { this._signatories.map(e => e.signature); });
   }
-
-  /**
-   * To sign transaction "asynchronously"
-   * @param {String} privKey Private key string
-   * @returns {String} a Promise that contains the transaction signature string
-   */
-  sign(privKey) {
-    // signing transaction require account address
-    // except signing for unbond txn, it requires validator address
-    const address = this.txnType === UBND_TXN_TYPE
-      ? gkeys.utils.crypto.getVaAddrByPrivKey(privKey)
-      : gkeys.utils.crypto.getAcAddrByPrivKey(privKey);
-    const getChainIdNSequence = () => {
-      const getAccount = () => {
-        return gkeys.utils.crypto.isVaAddress(address)
-          ? this.igc.account.getValidator(address)
-          : this.igc.account.getAccount(address)
-      }
-      return Promise.all([
-        this.igc.gallactic.getChainId(),
-        getAccount(address)
-      ]);
+  signSync (privKey) {
+    if (!privKey) {
+      throw new Error('Signing transaction require "privateKey" value. Please assign private key upon signing');
     }
-    const buildTxn = res => {
-      // res contains 2 promise (chainInfo and account details)
-      let chainRes = res[0], accRes = res[1];
-
-      if (!this.chainId && chainRes && chainRes.body && chainRes.body.result) {
-        this.chainId = chainRes.body.result.ChainId;
-      }
-
-      if (!this.sequence && accRes && accRes.body && accRes.body.result
-        && (accRes.body.result.Account || accRes.body.result.Validator)) {
-        this.sequence = (accRes.body.result.Account || accRes.body.result.Validator).sequence;
-      }
-      if (!this.chainId) {
-        throw new Error('Unable to retrieve Chain Id, please try again later');
-      }
-      if (typeof this.sequence !== 'number') {
-        throw new Error('Unable to retrieve sequence, please try again later');
-      }
-
-      this.txn = buildTxnByType(this._txn, this.txnType, { unit: this.txnUnit });
-      setTxnSequence(this.txn, this.txnType, this.sequence);
-      return;
-    };
-    const signTxn = () => {
-      try {
-        const signature = signByPrivKey(privKey, this.txn, this.txnType, this.chainId);
-        return Promise.resolve(signature);
-      } catch (e) {
-        throw e;
-      }
-    };
-
-    return getChainIdNSequence()
-      .then(buildTxn)
-      .then(signTxn)
-      .catch(e => {
-        throw e;
-      });
+    if (!this._chainId || (typeof this._sequence !== 'number' && !Array.isArray(this._sequence))) {
+      throw new Error('Signing synchronously require valid "chainId" and "sequence" value.')
+    }
+    this._setFrom(privKey);
+    // build txn
+    this._buildTxnByType();
+    // sign
+    this._setSignatories(privKey);
   }
-
-  broadcast() {
-    if (!this.chainId || typeof this.sequence !== 'number') {
-      throw new Error('Chain id or sequence is not defined. Unable to proceed');
-    }
-    if (!this.signatories || this.signatories.length === 0) {
-      throw new Error('Signatories is empty! Unable to proceed.');
-    }
-
+  broadcast () {
     let req = {
       bcMethod: txnMtd.broadcastTxn,
       params: {
-        chainId: this.chainId,
-        type: this.txnType,
+        chainId: this._chainId,
+        type: this._type,
         tx: this.txn,
-        signatories: this.signatories
+        signatories: this._signatories
       }
     };
 
     return this.conn.send(req);
   }
-
   /**
-   * To "BROADCAST" transaction based on "TRANSACTION TYPE" to the blockchain node
-   * @param {Array} signatories an array of object that contains public key and txn signature
+   * To broadcast "SEND transaction" to the blockchain node
+   * @param {String|Array} Private key string a String Array of Private Key
    * @returns {Object} a Promise that contains the transaction info
    */
-  signNBroadcast (privKey) {
-    let pubKey, signatories;
+  send (privKey) {
+    // set transaction type to SendTx disregarding previous value
+    this._type = SEND_TXN_TYPE;
+    return this.sign(privKey)
+      .then(() => { return this.broadcast() });
+  }
+  /**
+   * To broadcast "CALL transaction" to the blockchain node
+   * @param {String} Private key string a String Array of Private Key
+   * @returns {Object} a Promise that contains the transaction info
+   */
+  call (privKey) {
+    if (!this._gasLimit) throw new Error('Using call transaction require "gasLimit" field upon instantiating new transaction');
+    if (!this._data) throw new Error('Using call transaction require "data" field upon instantiating new transaction');
+    // set transaction type to SendTx disregarding previous value
+    this._type = CALL_TXN_TYPE;
+    return this.sign(privKey)
+      .then(() => { return this.broadcast() });
+  }
+  /**
+   * To broadcast "BOND transaction" to the blockchain node
+   * @param {String} Private key string a String Array of Private Key
+   * @returns {Object} a Promise that contains the transaction info
+   */
+  bond (privKey) {
+    if (!this._publicKey) throw new Error('Using bond transaction require "publicKey" field upon instantiating new transaction')
+    // set transaction type to SendTx disregarding previous value
+    this._type = BOND_TXN_TYPE;
+    return this.sign(privKey)
+      .then(() => { return this.broadcast() });
+  }
+  /**
+   * To broadcast "UNBOND transaction" to the blockchain node
+   * @param {String} Private key string a String Array of Private Key
+   * @returns {Object} a Promise that contains the transaction info
+   */
+  unbond (privKey) {
+    // set transaction type to SendTx disregarding previous value
+    this._type = UBND_TXN_TYPE;
+    return this.sign(privKey)
+      .then(() => { return this.broadcast() });
+  }
+  /**
+   * To broadcast "PERMISSION transaction" to the blockchain node
+   * @param {String} Private key string a String Array of Private Key
+   * @returns {Object} a Promise that contains the transaction info
+   */
+  permission (privKey) {
+    // set transaction type to SendTx disregarding previous value
+    this._type = PERM_TXN_TYPE;
+    return this.sign(privKey)
+      .then(() => { return this.broadcast() });
+  }
 
-    // get the tm pub key
-    pubKey = gkeys.utils.crypto.getTmPubKeyByPrivKey(privKey);
-    // base58 encode the pubkey
-    pubKey = gkeys.utils.crypto.bs58Encode(pubKey, 4) // TODO: replace 4 with gkeys.constant._PUB_KEY_ID
-    return this.sequence && this.chainId ? this.signSync(privKey) : this.sign(privKey)
-      .then(signature => {
-        signatories = [{ publicKey: pubKey, signature }]
 
-        validateSignatories(signatories);
+  /**
+   * =======================
+   * PRIVATE METHODS Section
+   * =======================
+   */
+  /**
+   * Helper function to set transaction type based on given type number
+   * @param {Number} txnType A transaction type number (1 - SEND_TXN, 2 - CALL_TXN)
+   * @returns {String} A string of transaction type signature
+   */
+  _getTxnType () {
+    if (this._typeId === SEND_TXN_ID) return SEND_TXN_TYPE;
+    else if (this._typeId === CALL_TXN_ID) return CALL_TXN_TYPE;
+    else if (this._typeId === BOND_TXN_ID) return BOND_TXN_TYPE;
+    else if (this._typeId === UBND_TXN_ID) return UBND_TXN_TYPE;
+    else if (this._typeId === PERM_TXN_ID) return PERM_TXN_TYPE;
+    else throw new Error('Unable to get Transaction sign of transaction type ' + this._typeId);
+  }
 
-        this.signatories = signatories;
+  /**
+   * Helper function to set from to `this` object
+   * @param  {String}  privKey  [Base 58 encoded string of Private key]
+   */
+  _setFrom (privKey) {
+    if (Array.isArray(privKey)) {
+      this._from = privKey.map(e => this._getAddress(e, this.type))
+    } else {
+      this._from = this._getAddress(privKey, this.type);
+    }
+  }
 
-        return this.broadcast();
+  /**
+   * Helper function to get address based on given private key
+   * @param   {String}  privKey  [private key string]
+   * @return  {String}  [a base 58 encoded address string]
+   */
+  _getAddress (privKey) {
+    return this._type === UBND_TXN_TYPE
+      ? gkeys.utils.crypto.getVaAddrByPrivKey(privKey)
+      : gkeys.utils.crypto.getAcAddrByPrivKey(privKey);
+  }
+
+  /**
+   * Helper function to set chainId to `this` object
+   */
+  _setChainId () {
+    return this._getChainId()
+      .then(chainId => this._chainId = chainId);
+  }
+
+  /**
+   * Helper function to get chain id of the node
+   * @return  {String}  [Chain Id string of the node]
+   */
+  _getChainId () {
+    let chainId;
+    return this.igc.gallactic.getChainId()
+      .then(res => {
+        if (res && res.body && res.body.result) {
+          chainId = res.body.result.ChainId;
+        }
+        if (!chainId) {
+          throw new Error('Unable to retrieve Chain Id, please try again later');
+        }
+        return chainId
       });
   }
 
   /**
-   * To broadcast "SEND transaction" to the blockchain node
-   * @param {Array} signatories an array of object that contains public key and txn signature
-   * @returns {Object} a Promise that contains the transaction info
+   * Helper function to set sequence to `this` object
    */
-  send(signatories = []) {
-    validateSignatories(signatories);
-    validateTxnType(this.txnType, SEND_TXN_TYPE);
-
-    this.signatories = signatories;
-
-    return this.broadcast();
+  _setSequence () {
+    let seqPromise;
+    if (Array.isArray(this._from)) {
+      seqPromise = Promise.all(this._from.map(e => this._getSequence(e)))
+    } else {
+      seqPromise = this._getSequence(this._from);
+    }
+    return seqPromise
+      .then(seq => this._sequence = seq)
   }
 
   /**
-   * To broadcast "CALL transaction" to the blockchain node
-   * @param {Array} signatories an array of object that contains public key and txn signature
-   * @returns {Object} a Promise that contains the transaction info
+   * Helper function to retrieve sequence number of an account
+   * @param   {String}  from  [base 58 encoded address string]
+   * @return  {Number}  [the transaction sequence of an account]
    */
-  call(signatories) {
-    validateSignatories(signatories);
-    validateTxnType(this.txnType, CALL_TXN_TYPE);
+  _getSequence (from) {
+    let sequence;
+    const getAccount = () => {
+      return gkeys.utils.crypto.isVaAddress(from)
+        ? this.igc.account.getValidator(from)
+        : this.igc.account.getAccount(from);
+    }
 
-    this.signatories = signatories;
+    return getAccount()
+      .then(res => {
+        if (res && res.body && res.body.result) {
+          sequence = (res.body.result.Account || res.body.result.Validator).sequence;
+        }
+        if (typeof sequence !== 'number') {
+          throw new Error('Unable to retrieve sequence, please try again later');
+        }
 
-    return this.broadcast();
+        // increment by 1 for next transaction
+        sequence += 1;
+        return sequence;
+      });
   }
 
   /**
-   * To broadcast "BOND transaction" to the blockchain node
-   * @param {Array} signatories an array of object that contains public key and txn signature
-   * @returns {Object} a Promise that contains the transaction info
+   * Helper function to set signatories to `this` object
+   * @param  {String}  privKey  [base 58 encoded string of private key]
    */
-  bond(signatories) {
-    validateSignatories(signatories);
-    validateTxnType(this.txnType, BOND_TXN_TYPE);
-
-    this.signatories = signatories;
-
-    return this.broadcast();
+  _setSignatories(privKey) {
+    let signatories = []
+    if (Array.isArray(privKey)) {
+      signatories = privKey.map(e => {
+        return this._getSignatories(e)
+      });
+    } else {
+      signatories.push(this._getSignatories(privKey))
+    }
+    this._signatories = signatories;
   }
 
   /**
-   * To broadcast "UNBOND transaction" to the blockchain node
-   * @param {Array} signatories an array of object that contains public key and txn signature
-   * @returns {Object} a Promise that contains the transaction info
+   * helper function to sign a transaction using private key and return the signatory object
+   * @param {String} privKey private key string
+   * @returns {Object} A signatory object that contains the signature and public key
    */
-  unbond(signatories) {
-    validateSignatories(signatories);
-    validateTxnType(this.txnType, UBND_TXN_TYPE);
-
-    this.signatories = signatories;
-
-    return this.broadcast();
-  }
-
-  /**
-   * To broadcast "PERMISSION transaction" to the blockchain node
-   * @param {Array} signatories an array of object that contains public key and txn signature
-   * @returns {Object} a Promise that contains the transaction info
-   */
-  permission(signatories) {
-    validateSignatories(signatories);
-    validateTxnType(this.txnType, PERM_TXN_TYPE);
-
-    this.signatories = signatories;
-
-    return this.broadcast();
-  }
-}
-
-/**
- * PRIVATE METHODS
- */
-
-/**
- * helper function to sign a transaction using private key and return a signature
- * @param {String} privKey private key string
- * @param {Object} txn Transaction object data
- * @param {String} txnType Transaction signature type (e.g. CallTx, SendTx)
- * @param {String} chainId Chain Id value
- * @returns {String} A signature of the transaction based on string format
- */
-function signByPrivKey (privKey, txn, txnType, chainId) {
-  let signObj;
-  signObj = JSON.stringify({
-    chainId: chainId,
-    type: txnType,
-    tx: txn
-  });
-
-  let data = {
-    // base 58 decode the private key and convert it to buffer
-    privKey: gkeys.utils.util.strToBuffer(
-      gkeys.utils.crypto.bs58Decode(privKey), 'hex'
-    ),
-    // message need to be keccak hashed before converted to buffer
-    message: gkeys.utils.util.strToBuffer(gkeys.utils.crypto.sha3(signObj))
-  };
-
-  return gkeys.utils.util.bytesToHexUpperString(
-    gkeys.utils.crypto.sign(data.message, data.privKey)
-  );
-}
-
-function buildTxnByType (tx, txnType, option) {
-  let txn = {};
-  if (txnType === SEND_TXN_TYPE) {
-    txn = {
-      senders: tx.from.map(e => setTxnFrom(e)),
-      receivers: tx.to.map(e => setTxnTo(e))
-    };
-  }
-  else if (txnType === CALL_TXN_TYPE) {
-    txn = {
-      caller: setTxnFrom(tx.from),
-      callee: setTxnTo(tx.to),
-      gasLimit: tx.gasLimit,
-      data: tx.data
-    };
-  }
-  else if (txnType === PERM_TXN_TYPE) {
-    txn = {
-      modifier: setTxnFrom(tx.from),
-      modified: setTxnTo(tx.to),
-      permissions: tx.permissions,
-      set: tx.set
-    };
-  }
-  else if (txnType === BOND_TXN_TYPE) {
-    txn = {
-      from: setTxnFrom(tx.from),
-      to: setTxnTo(tx.to),
-      public_key: tx.publicKey
-    }
-  }
-  else if (txnType === UBND_TXN_TYPE) {
-    txn = {
-      from: setTxnFrom(tx.from),
-      to: setTxnTo(tx.to)
-    }
-  }
-  else {
-    throw new Error('Cant build transaction object based on this type: ' + txnType);
-  }
-
-  return txn;
-
-  function setTxnFrom (from = {}, unit = option.unit) {
-    if (!from.address) {
-      throw new Error('Cant build transaction. from.address contains invalid address');
-    }
-    if (typeof from.amount === 'undefined' || typeof from.amount === 'null') {
-      throw new Error('Cant build transaction. from.amount contains invalid value')
-    }
-
-    const newFrom = {
-      address: from.address,
-      amount: conversion.toBoson(from.amount, unit).toNumber()
-    }
-    return newFrom;
-  }
-
-  function setTxnTo (to = {}, unit = option.unit) {
-    if (typeof to.amount === 'undefined' || typeof to.amount === 'null') {
-      throw new Error('Cant build transaction. to.amount contains invalid value')
-    }
-    const newTo = {
-      address: to.address,
-      amount: conversion.toBoson(to.amount, unit).toNumber()
-    };
-    return newTo;
-  }
-}
-
-/**
- * Helper function to set transaction sequence for send and call transaction
- * @param {Object} txn transaction object that contains necessary data
- * @param {Number} txnType transaction type identified e.g. txnType = 1
- * @param {Number} sequence a sequence number
- * @returns {undefined} no return value as this is just a helper to set txn sequence
- */
-function setTxnSequence (txn, txnType, sequence) {
-  if (Array.isArray(txn[getTxnFromKey(txnType)]) === true) {
-    txn[getTxnFromKey(txnType)].forEach(e => {
-      sequence = sequence + 1;
-      e.sequence = sequence;
+  _getSignatories (privKey) {
+    const signObj = JSON.stringify({
+      chainId: this._chainId,
+      type: this._type,
+      tx: this.txn
     });
-  }
-  else {
-    sequence = sequence + 1;
-    txn[getTxnFromKey(txnType)].sequence = sequence;
-  }
-}
 
-/**
- * Helper function to set transaction type based on given type number
- * @param {Number} txnType A transaction type number (1 - SEND_TXN, 2 - CALL_TXN)
- * @returns {String} A string of transaction type signature
- */
-function getTxnSign (txnType) {
-  if (txnType === SEND_TXN_ID) return SEND_TXN_TYPE;
-  else if (txnType === CALL_TXN_ID) return CALL_TXN_TYPE;
-  else if (txnType === BOND_TXN_ID) return BOND_TXN_TYPE;
-  else if (txnType === UBND_TXN_ID) return UBND_TXN_TYPE;
-  else if (txnType === PERM_TXN_ID) return PERM_TXN_TYPE;
-  else throw new Error('Unable to get Transaction sign of transaction type ' + txnType);
-}
-
-/**
- * Helper function to get transaction "from" key name based on txn type
- * @param {Number} txnType A transaction type number (1 - SEND_TXN, 2 - CALL_TXN)
- * @returns {String} A string of key name of "from" field
- */
-function getTxnFromKey (txnType) {
-  switch (txnType) {
-    case SEND_TXN_TYPE: return config.transactionFromKey.send;
-    case CALL_TXN_TYPE: return config.transactionFromKey.call;
-    case BOND_TXN_TYPE: return config.transactionFromKey.bond;
-    case UBND_TXN_TYPE: return config.transactionFromKey.ubnd;
-    case PERM_TXN_TYPE: return config.transactionFromKey.perm;
-    default: throw new Error('Unable to find from key by this type! unable to proceed');
+    const signature = gkeys.utils.util.bytesToHexUpperString(
+      gkeys.utils.crypto.sign(
+        // message need to be keccak hashed before converted to buffer
+        gkeys.utils.util.strToBuffer(gkeys.utils.crypto.sha3(signObj)), // message
+        // base 58 decode the private key and convert it to buffer
+        gkeys.utils.util.strToBuffer(gkeys.utils.crypto.bs58Decode(privKey), 'hex') // pvK buffer
+      )
+    );
+    return { signature, publicKey: gkeys.utils.crypto.getPubKeyByPrivKey(privKey) }
   }
-}
 
-/**
- * Helper function to help validate signatories object for broadcasting purpose
- * @param   {Array}  signatories  array of signatory object that contains pub key and signature
- */
-function validateSignatories (signatories) {
-  if (!Array.isArray(signatories) || signatories.length === 0) {
-    throw new Error('signatories with valid signature and pubKey is required for sending transaction')
-  }
-}
+  /**
+   * helper function to set txn to `this` object
+   */
+  _buildTxnByType () {
+    this.txn = {};
+    const from = this._from,
+      to = this._to,
+      sequence = this._sequence,
+      amountFrom = this._amount,
+      amountTo = this._amount,
+      unit = this._unit;
 
-/**
- * Helper function to help validate transaction type upon instanciate and function call
- * @param   {Array}  signatories  array of signatory object that contains pub key and signature
- */
-function validateTxnType (givenTxnType, expectedTxnType) {
-  if (givenTxnType !== expectedTxnType) {
-    throw new Error('Transaction was instantiated not for its purpose. Unable to proceed');
+    if (this._type === SEND_TXN_TYPE) {
+      this.txn = {
+        senders: Array.isArray(from)
+          ? from.map((e, i) => buildTxnFrom(e, sequence[i]))
+          : [buildTxnFrom(from, sequence)],
+        receivers: Array.isArray(to) ? to.map(e => buildTxnTo(e)) : [buildTxnTo(to)]
+      };
+    }
+    else if (this._type === CALL_TXN_TYPE) {
+      this.txn = {
+        caller: buildTxnFrom(from, sequence),
+        callee: buildTxnTo(to),
+        gasLimit: this._gasLimit,
+        data: this._data
+      };
+    }
+    else if (this._type === PERM_TXN_TYPE) {
+      this.txn = {
+        modifier: buildTxnFrom(from, sequence),
+        modified: buildTxnTo(to),
+        permissions: this._permissions,
+        set: this._set
+      };
+    }
+    else if (this._type === BOND_TXN_TYPE) {
+      this.txn = {
+        from: buildTxnFrom(from, sequence),
+        to: buildTxnTo(to),
+        public_key: this._publicKey
+      }
+    }
+    else if (this._type === UBND_TXN_TYPE) {
+      this.txn = {
+        from: buildTxnFrom(from, sequence),
+        to: buildTxnTo(to)
+      }
+    }
+    else {
+      throw new Error('Cant build transaction object based on this type: ' + this._type);
+    }
+
+    function buildTxnFrom (from, seq) {
+      if (!from && gkeys.utils.crypto.isAddress(from)) {
+        throw new Error('Cant build transaction. "from" contains invalid address');
+      }
+      if (typeof seq !== 'number') {
+        throw new Error('Cant build transaction. "sequence" contains invalid value')
+      }
+      if (typeof amountFrom !== 'number') {
+        throw new Error('Cant build transaction. "amount" contains invalid value')
+      }
+
+      const newFrom = {
+        address: from,
+        amount: conversion.toBoson(amountFrom, unit).toNumber(),
+        sequence: seq
+      }
+      return newFrom;
+    }
+
+    function buildTxnTo (to) {
+      if (!to && gkeys.utils.crypto.isAddress(to)) {
+        throw new Error('Cant build transaction. to.amount contains invalid value')
+      }
+      if (typeof amountTo !== 'number') {
+        throw new Error('Cant build transaction. from.amount contains invalid value')
+      }
+      const newTo = {
+        address: to,
+        amount: conversion.toBoson(amountTo, unit).toNumber()
+      };
+      return newTo;
+    }
   }
 }
